# docs/swagger.yaml
basePath: /
definitions:
  models.BookingDTO:
    properties:
      employee_id:
        type: integer
      end_time:
        type: string
      num_attendees:
        type: integer
      room_id:
        type: integer
      start_time:
        type: string
    type: object
...
host: localhost:9010
info:
  contact: {}
  description: API documentation for Meeting Room Booking system
  title: Meeting Room Booking API
  version: "1.0"
paths:
  /booking/{id}:
    delete:
      consumes:
      - application/json
      description: Allows an authenticated employee to delete their booking
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: integer
      - description: Deleted booking details
        in: body
        name: booking
        required: true
        schema:
          $ref: '#/definitions/models.BookingDTO'
      produces:
      - application/json
      responses:
        "400":
          description: Invalid Employee ID or JSON input
          schema:
            type: string
        "401":
          description: Unauthorized (not logged in)
          schema:
            type: string
        "403":
          description: Forbidden (trying to update another employee's booking)
          schema:
            type: string
        "404":
          description: Booking not found
          schema:
            type: string
        "500":
          description: Failed to delete booking
          schema:
            type: string
      summary: Delete existing booking details
      tags:
      - Bookings
    put:
      consumes:
      - application/json
      description: Allows an authenticated employee to update their booking
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated booking details
        in: body
        name: booking
        required: true
        schema:
          $ref: '#/definitions/models.BookingDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.BookingDTO'
        "400":
          description: Invalid Employee ID or JSON input
          schema:
            type: string
        "401":
          description: Unauthorized (not logged in)
          schema:
            type: string
        "403":
          description: Forbidden (trying to update another employee's booking)
          schema:
            type: string
        "404":
          description: Booking not found
          schema:
            type: string
        "500":
          description: Failed to update booking
          schema:
            type: string
      summary: Update existing booking details
      tags:
      - Bookings
  /bookings:
    get:
      description: Retrieves all bookings along with all details
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.BookingDTO'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get list of all bookings
      tags:
      - Bookings
    post:
      consumes:
      - application/json
      description: Creates a new booking if the time and capacity constraints are
        satisfied. Sends confirmation email and adds Google Calendar event if linked.
      parameters:
      - description: Booking request data
        in: body
        name: booking
        required: true
        schema:
          $ref: '#/definitions/models.BookingDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.BookingDTO'
        "400":
          description: Invalid input or time conflict
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "409":
          description: Booking time conflict
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Create a new booking
      tags:
      - Bookings
  /bookings/{id}:
    get:
      description: Retrieves booking details by ID including employee and room info
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.BookingDTO'
        "400":
          description: Invalid ID
          schema:
            type: string
        "404":
          description: Booking not found
          schema:
            type: string
        "500":
          description: Error marshalling data
          schema:
            type: string
      summary: Get booking by booking ID
      tags:
      - Bookings
...
swagger: "2.0"